Make updatesModal a component
