+++ client/src/app/main-menu/main-menu.module.scss
@@ -236,31 +236,6 @@ nav {
 	gap: 1vh;
 }
 
-.updatesList {
-	display: flex;
-	flex-direction: column-reverse;
-	align-items: center;
-	overflow-y: scroll;
-	color: var(--text-light);
-	gap: 10px;
-	max-height: 40vh;
-	height: 100%;
-
-	a {
-		color: cyan;
-	}
-}
-
-.updateItem {
-	width: 100%;
-	padding: 1%;
-	text-align: center;
-}
-
-.updateSeperator {
-	width: 90%;
-}
-
 /* DESKTOP */
 @media (min-width: 80vh) {
 	nav {
