+++ client/src/app/main-menu/main-menu.tsx
@@ -11,6 +11,7 @@ import {useEffect, useRef, useState} from 'react'
 import AlertModal from 'components/alert-modal'
 import {toHTML} from 'discord-markdown'
 import DOMPurify from 'dompurify'
+import UpdatesModal from 'components/updates'
 
 type Props = {
 	setMenuSection: (section: string) => void
@@ -28,54 +29,15 @@ function MainMenu({setMenuSection}: Props) {
 	const updates = useSelector(getUpdates)
 	const [updatesOpen, setUpdatesOpen] = useState<boolean>(true)
 	const latestUpdateView = localStorage.getItem('latestUpdateView')
-	const latestUpdateElement = useRef<HTMLLIElement>(null)
-	useEffect(() => {
-		latestUpdateElement.current?.scrollIntoView({
-			behavior: 'instant',
-			block: 'start',
-		})
-	})
 
 	const welcomeMessage = playerDeck.name === 'Starter Deck' ? 'Welcome' : 'Welcome Back'
 
 	return (
 		<>
 			{!latestUpdateView ||
-			parseInt(updates['timestamps'] ? updates['timestamps'][0] : '0') > parseInt(latestUpdateView) ? (
-				<AlertModal
-					setOpen={updatesOpen}
-					onClose={() => {
-						setUpdatesOpen(false)
-						localStorage.setItem('latestUpdateView', (new Date().valueOf() / 1000).toFixed())
-					}}
-					cancelText="Close"
-					title="Latest updates"
-					action={() => {}}
-					description={
-						<ul className={css.updatesList}>
-							{updates['updates'] ? (
-								updates['updates'].map((text, i) => {
-									return (
-										<>
-											<li
-												className={css.updateItem}
-												key={i + 1}
-												dangerouslySetInnerHTML={{__html: DOMPurify.sanitize(toHTML(text))}}
-												ref={i === 0? latestUpdateElement: undefined}
-											/>
-											<hr key={-i} className={css.updateSeperator} />
-										</>
-									)
-								})
-							) : (
-								<li className={css.updateItem}>Failed to load updates</li>
-							)}
-							<li key={20} className={css.updateItem}>
-								For more updates, visit the HC-TCG discord.
-							</li>
-						</ul>
-					}
-				/>
+			parseInt(updates['timestamps'] ? updates['timestamps'][0] : '0') >
+				parseInt(latestUpdateView) ? (
+				<UpdatesModal updatesOpen={updatesOpen} setUpdatesOpen={setUpdatesOpen} />
 			) : (
 				<></>
 			)}
